+++ src/components/pages/apps/matcalc/apps/sankey/sankey-dialog.tsx
@@ -2,7 +2,12 @@ import { TEXT_OK } from '@/consts'
 import { OKCancelDialog, type IModalProps } from '@/dialogs/ok-cancel-dialog'
 
 import { useDialogs } from '@/components/dialogs/dialogs'
+import { PropRow } from '@/components/dialogs/prop-row'
+import { SettingsAccordionItem } from '@/components/dialogs/settings/settings-dialog'
+import { ScrollAccordion } from '@/components/shadcn/ui/themed/v2/accordion'
 import { BaseDataFrame } from '@/lib/dataframe/base-dataframe'
+import { useState } from 'react'
+import { DFColSelect } from '../../df-col-select'
 import { useCurrentSheets } from '../../history/history-provider/history-contexts'
 import { DFToSankeyGraph } from './sankey-parser'
 import { ISankeyPlot } from './sankey-provider'
@@ -28,9 +33,33 @@ export function SankeyDialog({
   const { sheet } = useCurrentSheets()
   const { open: openDialog } = useDialogs()
 
+  const [sourceCol, setSourceCol] = useState('source')
+  const [sourceColCol, setSourceColCol] = useState('source column')
+  const [targetCol, setTargetCol] = useState('target')
+  const [targetColCol, setTargetColCol] = useState('target column')
+  const [valueCol, setValueCol] = useState('value')
+
+  const [sourceColorCol, setSourceColorCol] = useState('source color')
+  const [targetColorCol, setTargetColorCol] = useState('target color')
+  const [linkColorCol, setLinkColorCol] = useState('link color')
+
+  const [accordionValues, setAccordionValues] = useState<string[]>([
+    'required',
+    'optional',
+  ])
+
   function makeSankeyPlot() {
     try {
-      const plot = DFToSankeyGraph(sheet as BaseDataFrame)
+      const plot = DFToSankeyGraph(sheet as BaseDataFrame, {
+        sourceCol,
+        sourceColCol,
+        targetCol,
+        targetColCol,
+        valueCol,
+        sourceColorCol,
+        targetColorCol,
+        linkColorCol,
+      })
 
       onResponse?.(TEXT_OK, plot)
     } catch (e) {
@@ -56,6 +85,74 @@ export function SankeyDialog({
         }
       }}
       //contentVariant="glass"
-    ></OKCancelDialog>
+    >
+      <ScrollAccordion
+        value={accordionValues}
+        onValueChange={(v) => setAccordionValues(v as string[])}
+        variant="settings"
+        className="h-64"
+      >
+        <SettingsAccordionItem title={`Required headings`} value="required">
+          <PropRow title="Source">
+            <DFColSelect
+              df={sheet as BaseDataFrame}
+              value={sourceCol}
+              onChange={({ value }) => setSourceCol(value)}
+            />
+          </PropRow>
+          <PropRow title="Source column">
+            <DFColSelect
+              df={sheet as BaseDataFrame}
+              value={sourceColCol}
+              onChange={({ value }) => setSourceColCol(value)}
+            />
+          </PropRow>
+          <PropRow title="Target">
+            <DFColSelect
+              df={sheet as BaseDataFrame}
+              value={targetCol}
+              onChange={({ value }) => setTargetCol(value)}
+            />
+          </PropRow>
+          <PropRow title="Target column">
+            <DFColSelect
+              df={sheet as BaseDataFrame}
+              value={targetColCol}
+              onChange={({ value }) => setTargetColCol(value)}
+            />
+          </PropRow>
+          <PropRow title="Value">
+            <DFColSelect
+              df={sheet as BaseDataFrame}
+              value={valueCol}
+              onChange={({ value }) => setValueCol(value)}
+            />
+          </PropRow>
+        </SettingsAccordionItem>
+        <SettingsAccordionItem title={`Optional headings`} value="optional">
+          <PropRow title="Source color">
+            <DFColSelect
+              df={sheet as BaseDataFrame}
+              value={sourceColorCol}
+              onChange={({ value }) => setSourceColorCol(value)}
+            />
+          </PropRow>
+          <PropRow title="Target color">
+            <DFColSelect
+              df={sheet as BaseDataFrame}
+              value={targetColorCol}
+              onChange={({ value }) => setTargetColorCol(value)}
+            />
+          </PropRow>
+          <PropRow title="Link color">
+            <DFColSelect
+              df={sheet as BaseDataFrame}
+              value={linkColorCol}
+              onChange={({ value }) => setLinkColorCol(value)}
+            />
+          </PropRow>
+        </SettingsAccordionItem>
+      </ScrollAccordion>
+    </OKCancelDialog>
   )
 }